requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==24.2.0
click==8.1.7
Django==5.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
environs==10.3.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
marshmallow==3.20.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
uritemplate==4.1.1
